+++ include/nstd/events/def.h
@@ -1,0 +1,9 @@
+#ifndef NSTD_EVENTS_DEF_H_INCLUDED
+#define NSTD_EVENTS_DEF_H_INCLUDED
+#include "../core/def.h"
+#include "../nstd.h"
+
+/// Represents a button ID.
+typedef NSTDUInt32 NSTDButtonID;
+
+#endif