+++ src/events/def.rs
@@ -1,0 +1,4 @@
+//! Common types that are used by the event system.
+
+/// Represents a button ID.
+pub type NSTDButtonID = u32;